+++ flow/flow_ebos_blackoil.cpp
@@ -16,7 +16,10 @@
 */
 #include "config.h"
 
-#include <opm/simulators/flow_ebos_solvent.hpp>
+// Define making clear that the simulator supports AMG
+#define FLOW_SUPPORT_AMG 1
+
+#include <flow/flow_ebos_blackoil.hpp>
 
 #include <opm/material/common/ResetLocale.hpp>
 #include <opm/grid/CpGrid.hpp>
@@ -29,37 +32,30 @@
 #include <dune/common/parallel/mpihelper.hh>
 #endif
 
-namespace Ewoms {
-namespace Properties {
-NEW_TYPE_TAG(EclFlowSolventProblem, INHERITS_FROM(EclFlowProblem));
-SET_BOOL_PROP(EclFlowSolventProblem, EnableSolvent, true);
-}}
-
 namespace Opm {
-void flowEbosSolventSetDeck(Deck &deck, EclipseState& eclState, Schedule& schedule, SummaryConfig& summaryConfig)
+
+void flowEbosBlackoilSetDeck(Deck &deck, EclipseState& eclState, Schedule& schedule, SummaryConfig& summaryConfig)
 {
-    typedef TTAG(EclFlowSolventProblem) TypeTag;
+    typedef TTAG(EclFlowProblem) TypeTag;
     typedef GET_PROP_TYPE(TypeTag, Vanguard) Vanguard;
 
     Vanguard::setExternalDeck(&deck, &eclState, &schedule, &summaryConfig);
 }
 
-
 // ----------------- Main program -----------------
-int flowEbosSolventMain(int argc, char** argv)
+int flowEbosBlackoilMain(int argc, char** argv)
 {
     // we always want to use the default locale, and thus spare us the trouble
     // with incorrect locale settings.
     Opm::resetLocale();
 
-    // initialize MPI, finalize is done automatically on exit
 #if HAVE_DUNE_FEM
     Dune::Fem::MPIManager::initialize(argc, argv);
 #else
-    Dune::MPIHelper::instance(argc, argv).rank();
+    Dune::MPIHelper::instance(argc, argv);
 #endif
 
-    Opm::FlowMainEbos<TTAG(EclFlowSolventProblem)> mainfunc;
+    Opm::FlowMainEbos<TTAG(EclFlowProblem)> mainfunc;
     return mainfunc.execute(argc, argv);
 }
