+++ flow/flow_ebos_energy.cpp
@@ -16,10 +16,7 @@
 */
 #include "config.h"
 
-// Define making clear that the simulator supports AMG
-#define FLOW_SUPPORT_AMG 1
-
-#include <opm/simulators/flow_ebos_blackoil.hpp>
+#include <flow/flow_ebos_energy.hpp>
 
 #include <opm/material/common/ResetLocale.hpp>
 #include <opm/grid/CpGrid.hpp>
@@ -32,30 +29,36 @@
 #include <dune/common/parallel/mpihelper.hh>
 #endif
 
-namespace Opm {
+namespace Ewoms {
+namespace Properties {
+NEW_TYPE_TAG(EclFlowEnergyProblem, INHERITS_FROM(EclFlowProblem));
+SET_BOOL_PROP(EclFlowEnergyProblem, EnableEnergy, true);
+}}
 
-void flowEbosBlackoilSetDeck(Deck &deck, EclipseState& eclState, Schedule& schedule, SummaryConfig& summaryConfig)
+namespace Opm {
+void flowEbosEnergySetDeck(Deck &deck, EclipseState& eclState, Schedule& schedule, SummaryConfig& summaryConfig)
 {
-    typedef TTAG(EclFlowProblem) TypeTag;
+    typedef TTAG(EclFlowEnergyProblem) TypeTag;
     typedef GET_PROP_TYPE(TypeTag, Vanguard) Vanguard;
 
     Vanguard::setExternalDeck(&deck, &eclState, &schedule, &summaryConfig);
 }
 
 // ----------------- Main program -----------------
-int flowEbosBlackoilMain(int argc, char** argv)
+int flowEbosEnergyMain(int argc, char** argv)
 {
     // we always want to use the default locale, and thus spare us the trouble
     // with incorrect locale settings.
     Opm::resetLocale();
 
+    // initialize MPI, finalize is done automatically on exit
 #if HAVE_DUNE_FEM
     Dune::Fem::MPIManager::initialize(argc, argv);
 #else
-    Dune::MPIHelper::instance(argc, argv);
+    Dune::MPIHelper::instance(argc, argv).rank();
 #endif
 
-    Opm::FlowMainEbos<TTAG(EclFlowProblem)> mainfunc;
+    Opm::FlowMainEbos<TTAG(EclFlowEnergyProblem)> mainfunc;
     return mainfunc.execute(argc, argv);
 }
